Move a ViewAction to the custom package, to keep the project clean
