+++ library/src/main/java/com/schibsted/spain/barista/BaristaClickActions.java
@@ -11,7 +11,7 @@ import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.Espresso.pressBack;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
-import static com.schibsted.spain.barista.NestedEnabledScrollToAction.scrollTo;
+import static com.schibsted.spain.barista.custom.NestedEnabledScrollToAction.scrollTo;
 import static com.schibsted.spain.barista.custom.DisplayedMatchers.displayedAnd;
 
 public class BaristaClickActions {
